\name{dynamo-package}
\alias{dynamo-package}
\docType{package}
\title{DYNAmic MOdelling}
\description{The dynamo package provides routines for the analysis of time series data.}
\details{
\tabular{ll}{
Package: \tab dynamo\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2014-12-23\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
} }
\author{
Christian Brownlees 

Maintainer: Christian Brownlees 
}
\keyword{package}
