\name{predict.dm}
\alias{predict.dm}
\title{ Forecast dynamic models }
\description{ Forecasts from models fitted via dm }
\usage{ predict( obejct , n.ahead=NULL) }
\arguments{
	\item{object}{an object of class "dm".}
	\item{n.ahead}{number of periods ahead}
}
\details{}
\value{
	Depending on the class 
}
\references{}
\author{Christian Brownlees}
\examples{

garch11

}
\keyword{timeseries}
